requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
python-dotenv==1.0.1
Werkzeug==3.0.3
